requirements.txt
Django==5.2.3
whitenoise==6.7.0
dj-database-url==2.3.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
